Resolved crash.
Adding a station with a name that already exists resulted in
the editor crashing. This was due to a Window state issue with
the modal dialog. Leaving the modal dialog in a non-destroyed
state allows the editor to continue as expected. It was also
found that the call to Destroy() was actually unnecessary. As
such, all calls in the addStation() method were removed.

Also, resolved a memory leak issue when shutting down.